backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0

# Data handling
pandas==2.1.4
numpy==1.26.2
beautifulsoup4==4.12.2

# YouTube and search
youtube-transcript-api==0.6.1
duckduckgo-search==3.9.6
yfinance==0.2.28

# LLM and agents
openai>=1.86.0
pydantic>=2.7.4
pydantic-core>=2.18.2
